chore(mocha): exclude node_modules from tests
I had an awkward Mocha config which I have clarified with this commit.
Rather than using an unintuitive glob in the npm script, I have added
Mocha config options within a new section in `package.json` to ignore
the `node_modules` folder and to search recursively. This leaves the
file parsing down to the npm script to allow for scenarios, e.g. just
unit tests or just integration tests. These are done by passing flags to
the command, the deafault for which I have set to files ending in
`.spec.js` and `.test.js`.