feat(apex): update Apex squad randomizer
New look for the Apex squad randomizer which mimics the in-game teammte
display. Rather than displaying all legends in a single MessageEmbed,
each legend now has its own embed with thumbnail and colours to match
the game.

The images live in the `.gitignore`-ed `Resources` directory (to avoid
uploading binaries to GitHub and potential copyright/'not fair use'
issues), so will need to be copied to the host server manually.